requirements.txt
aiohttp==3.9.3
aioselenium==0.0.1
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
DAWG-Python==0.7.2
Django==4.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-ckeditor5==0.0.4
django-cors==0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-dotenv==1.4.2
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-nested-admin==4.0.2
django-tables2==2.7.0
djangorestframework==3.15.1
docopt==0.6.2
drf-spectacular==0.27.1
drf-yasg==1.21.7
facebook-sdk==3.1.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pymorphy2-dicts-ru==2.4.417127.4579844
python-decouple==3.8
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
